backend/requirements.txt
# =============================================================================
# backend/requirements.txt — Lambda + Test Dependencies
# =============================================================================
# Production dependencies (included in Lambda deployment package):
#   boto3          — AWS SDK for S3, DynamoDB, SES, Bedrock
#   strands-agents — AWS open-source agent framework (tools + orchestration)
#
# Development/test dependencies (NOT included in Lambda package):
#   pytest             — test runner
#   pytest-html        — HTML test report generation
#   pytest-cov         — test coverage reporting
#   moto               — AWS service mocking for unit tests (S3, DynamoDB, SES, Bedrock)
#
# To install production deps only (Lambda layer):
#   pip install boto3 strands-agents strands-agents-tools
#
# To install all deps (local development + testing):
#   pip install -r requirements.txt
# =============================================================================

# --- Production ---
boto3>=1.34.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# --- Testing ---
pytest>=8.0.0
pytest-html>=4.1.0
pytest-cov>=5.0.0
moto[s3,dynamodb,ses,bedrock]>=5.0.0
